pyproject.toml
[project]
name = "clothing-ecommerce"
version = "0.1.0"
description = "Kafka driven clothing ecommerce bootstrap"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Abhishek Hada" }]
dependencies = [
  "kafka-python==2.0.2",
  "tinydb==4.8.2",
  "Faker==33.0.0",
  "fastapi==0.115.4",
  "uvicorn[standard]==0.30.6",
  "python-dotenv==1.0.1",
  "pydantic==2.9.2",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.coverage.run]
branch = true
source = [
  "cart_service", 
  "order_service", 
  "payment_service", 
  "inventory_service", 
  "shipping_service", 
  "notification_service", 
  "common"
  ]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 80

[project.optional-dependencies]
dev = [
  "httpx==0.27.2",
  "pytest==8.3.3",
  "pytest-asyncio==0.23.8",
  "pytest-cov==5.0.0",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov --cov-report=term-missing"
testpaths = [
  "common/tests", 
  "inventory_service/tests",
  "cart_service/tests"
  ]
